--- server/MainProject/src/main/java/com/main/MainProject/cart/repository/CartRepository.java
@@ -1,7 +1,0 @@
-package com.main.MainProject.cart.repository;
-
-import com.main.MainProject.cart.entity.Cart;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CartRepository extends JpaRepository<Cart, Long> {
-}
