--- server/MainProject/src/main/java/com/main/MainProject/dto/PageInfo.java
@@ -1,13 +1,0 @@
-package com.main.MainProject.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class PageInfo {
-    private int page;
-    private int size;
-    private long totalElements;
-    private int totalPages;
-}
