--- server/MainProject/src/main/java/com/main/MainProject/dto/SingleResponseDto.java
@@ -1,11 +1,0 @@
-package com.main.MainProject.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class SingleResponseDto<T> {
-    private T data;
-
-}
