--- server/MainProject/src/main/java/com/main/MainProject/exception/BusinessLogicException.java
@@ -1,13 +1,0 @@
-package com.main.MainProject.exception;
-
-import lombok.Getter;
-
-public class BusinessLogicException extends RuntimeException{
-    @Getter
-    private ExceptionCode exceptionCode;
-
-    public BusinessLogicException(ExceptionCode exceptionCode) {
-        super(exceptionCode.getMessage());
-        this.exceptionCode = exceptionCode;
-    }
-}
