+++ server/MainProject/src/main/java/com/main/MainProject/order/OrderDto.java
@@ -1,0 +1,16 @@
+package com.main.MainProject.order;
+
+public class OrderDto {
+
+    public static class orderPost{
+
+    }
+
+    public static class patch{
+
+    }
+
+    public static class response{
+
+    }
+}
