+++ server/MainProject/src/main/java/com/main/MainProject/order/OrderMapper.java
@@ -1,0 +1,8 @@
+package com.main.MainProject.order;
+
+import org.mapstruct.Mapper;
+import org.mapstruct.ReportingPolicy;
+
+@Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
+public interface OrderMapper {
+}
