+++ server/MainProject/src/main/java/com/main/MainProject/order/OrderRepository.java
@@ -1,6 +1,5 @@
-package com.main.MainProject.order.repository;
+package com.main.MainProject.order;
 
-import com.main.MainProject.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 public interface OrderRepository extends JpaRepository<Order, Long> {
 }
