+++ server/MainProject/src/main/java/com/main/MainProject/order/OrderService.java
@@ -1,0 +1,4 @@
+package com.main.MainProject.order;
+
+public class OrderService {
+}
