--- server/MainProject/src/main/java/com/main/MainProject/temporary/AddressRepository.java
@@ -1,6 +1,0 @@
-package com.main.MainProject.temporary;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AddressRepository extends JpaRepository<Address, Long> {
-}
