--- server/MainProject/src/main/java/com/main/MainProject/temporary/Member.java
@@ -1,30 +1,0 @@
-package com.main.MainProject.temporary;
-
-import com.main.MainProject.cart.entity.Cart;
-import com.main.MainProject.order.entity.Order;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-public class Member {
-    @Id
-    private Long memberId;
-
-    private String name;
-
-    private String phone;
-
-    private String nickname;
-
-
-//    @OneToOne(mappedBy = "member", cascade = CascadeType.ALL, orphanRemoval = true)
-//    private Address address;
-
-//    @OneToOne(mappedBy = "member", cascade = CascadeType.ALL, orphanRemoval = true)
-//    private Cart cart;
-
-//    @OneToMany(mappedBy = "member", cascade = CascadeType.PERSIST, orphanRemoval = true)
-//    private List<Order> orderList = new ArrayList<>();
-}
